requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
